requirements.txt
pytesseract
beautifulsoup4
requests
google-auth
google-api-python-client
